requirements.txt
# Remove or replace lines like these with the actual package names and versions
# aext-assistant @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_5blio7a88h/croot/aext-assistant_1717062156186/work
# Replace with something like:
# aext-assistant==1.0.0

dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.0.0
django-filter==24.3.0
djangorestframework==3.15.2
gunicorn==23.0.0
psycopg2-binary==2.9.7
